requirements.txt
flask==3.0.0
python-dotenv==1.0.0
baml-py==0.89.0
requests==2.31.0
Flask-CORS
langchain
langchain-core
langchain-google-genai
openai>=1.0.0
boto3
google-generativeai
uuid